requirements.txt
Django
azure-identity
azure-keyvault
azure-storage-blob
azure-core
azure-common